flask_app/requirements.txt
Flask
boto3
requests
Pillow
sagemaker
